+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.1 (2017-11-05)
+
+- Ensure that `plugged` obj is also available in overriden constructors
+- Plugins now can have an `enabled` attribute (either a method or a property)
+  which determines whether the plugin should be initialized or not.
+
 ## 1.0.0 (2017-02-25)
 
 - Use lodash instead of underscore.